# docs/swagger.yaml
basePath: /
definitions:
  controllers.Response:
    properties:
      message:
        type: string
      page_info: {}
      results: {}
      success:
        type: boolean
    type: object
  models.GetAllMovie:
    properties:
      genre:
        example: Action
        type: string
      id:
        type: integer
      image:
        example: Spiderman.jpg
        type: string
      tittle:
        example: Spiderman
        type: string
    type: object
  models.Movie_body:
    properties:
      actors:
        type: string
      author:
        type: string
      duration:
        type: string
      genre:
        example: Action
        type: string
      id:
        type: integer
      image:
        example: Spiderman.jpg
        type: string
      release_date:
        type: string
      synopsis:
        example: film action universal
        type: string
      tag:
        description: |-
          Release_date time.Time `json:"release_date" form:"release_date"`
          Duration     time.Time `json:"duration" form:"duration"`
        type: string
      tittle:
        example: Spiderman
        type: string
    type: object
  models.MoviesCinema:
    properties:
      cinema:
        type: string
      cinema_date:
        type: string
      cinema_location:
        type: string
      cinema_time:
        type: string
      genre:
        type: string
      id:
        type: integer
      image:
        type: string
      tittle:
        type: string
    type: object
  models.MoviesNoTag:
    properties:
      actors:
        type: string
      author:
        type: string
      duration:
        type: string
      genre:
        example: Action
        type: string
      id:
        type: integer
      image:
        example: Spiderman.jpg
        type: string
      release_date:
        type: string
      synopsis:
        example: film action universal
        type: string
      tittle:
        example: Spiderman
        type: string
    type: object
  models.MoviesbyTag:
    properties:
      actors:
        type: string
      author:
        type: string
      genre:
        example: Action
        type: string
      id:
        type: integer
      image:
        example: Spiderman.jpg
        type: string
      synopsis:
        example: film action universal
        type: string
      tag:
        description: |-
          Release_date time.Time `json:"release_date" form:"release_date"`
          Duration     time.Time `json:"duration" form:"duration"`
        type: string
      tittle:
        example: Spiderman
        type: string
    type: object
  models.OrderBody:
    properties:
      cinema_id:
        type: integer
      cinema_name:
        type: string
      date:
        type: string
      genre:
        type: string
      id:
        type: integer
      image:
        type: string
      location:
        type: string
      movie_id:
        type: integer
      payment_id:
        type: integer
      profile_id:
        type: integer
      quantity:
        type: integer
      seat_id:
        type: integer
      time:
        type: string
      title:
        type: string
      total_price:
        type: integer
    type: object
  models.PointProfile:
    properties:
      email:
        example: salah@mail.com
        type: string
      first_name:
        example: Salah
        type: string
      id:
        type: integer
      image:
        example: salah.jpg
        type: string
      last_name:
        example: Alaudin
        type: string
      password:
        example: Salah1!
        type: string
      phone_number:
        type: string
      point:
        type: integer
    type: object
  models.Profile:
    properties:
      email:
        example: salah@mail.com
        type: string
      first_name:
        example: Salah
        type: string
      id:
        type: integer
      image:
        example: salah.jpg
        type: string
      last_name:
        example: Alaudin
        type: string
      password:
        example: Salah1!
        type: string
      phone_number:
        type: string
    type: object
info:
  contact: {}
  description: Profile server
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Login Account
      parameters:
      - description: Input Email
        in: formData
        name: email
        required: true
        type: string
      - description: Input Password
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.Response'
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Registrasi Account
      parameters:
      - description: Input Email
        in: formData
        name: email
        required: true
        type: string
      - description: Input Password
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.Response'
      tags:
      - Auth
  /movies:
    get:
      consumes:
      - application/json
      description: Get All Movie
      parameters:
      - description: Search Movie
        in: query
        name: search
        type: string
      - description: Page Movie
        in: query
        name: page
        type: integer
      - description: Limit Movie
        in: query
        name: limit
        type: integer
      - description: Sort Movie
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controllers.Response'
            - properties:
                results:
                  items:
                    $ref: '#/definitions/models.GetAllMovie'
                  type: array
              type: object
      summary: Movies
      tags:
      - Movies
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Add Movies
      parameters:
      - description: Id Movie
        in: path
        name: id
        required: true
        type: integer
      - description: Edit tittle
        in: formData
        name: tittle
        required: true
        type: string
      - description: Update genre
        in: formData
        name: genre
        required: true
        type: string
      - description: Update Image
        in: formData
        name: Image
        required: true
        type: file
      - description: Update Synopsis
        in: formData
        name: synopsis
        required: true
        type: string
      - description: Update Author
        in: formData
        name: author
        required: true
        type: string
      - description: Update Actors
        in: formData
        name: actors
        required: true
        type: string
      - description: Update Realease Date
        in: formData
        name: release_date
        required: true
        type: string
      - description: Update Duration
        in: formData
        name: duration
        required: true
        type: string
      - description: Update Tag
        in: formData
        name: tag
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controllers.Response'
            - properties:
                results:
                  $ref: '#/definitions/models.MoviesbyTag'
              type: object
      tags:
      - Movies
  /movies/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Movies
      parameters:
      - description: Delete Movie
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controllers.Response'
            - properties:
                results:
                  $ref: '#/definitions/models.MoviesbyTag'
              type: object
      tags:
      - Movies
    get:
      consumes:
      - application/json
      description: Detail Movies
      parameters:
      - description: Detail Movie
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controllers.Response'
            - properties:
                results:
                  $ref: '#/definitions/models.MoviesNoTag'
              type: object
      tags:
      - Movies
    patch:
      consumes:
      - application/x-www-form-urlencoded
      description: Edit Movies
      parameters:
      - description: Id Movie
        in: path
        name: id
        required: true
        type: integer
      - description: Edit tittle
        in: formData
        name: tittle
        required: true
        type: string
      - description: Update genre
        in: formData
        name: genre
        required: true
        type: string
      - description: Update Image
        in: formData
        name: Image
        required: true
        type: file
      - description: Update Synopsis
        in: formData
        name: synopsis
        required: true
        type: string
      - description: Update Author
        in: formData
        name: author
        required: true
        type: string
      - description: Update Actors
        in: formData
        name: actors
        required: true
        type: string
      - description: Update Realease Date
        in: formData
        name: release_date
        required: true
        type: string
      - description: Update Duration
        in: formData
        name: duration
        required: true
        type: string
      - description: Update Tag
        in: formData
        name: tag
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controllers.Response'
            - properties:
                results:
                  $ref: '#/definitions/models.Movie_body'
              type: object
      tags:
      - Movies
  /orders:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Add Order tikers
      parameters:
      - in: formData
        name: cinema_id
        type: integer
      - in: formData
        name: cinema_name
        type: string
      - in: formData
        name: date
        type: string
      - in: formData
        name: genre
        type: string
      - in: formData
        name: id
        type: integer
      - in: formData
        name: image
        type: string
      - in: formData
        name: location
        type: string
      - in: formData
        name: movie_id
        type: integer
      - in: formData
        name: payment_id
        type: integer
      - in: formData
        name: profile_id
        type: integer
      - in: formData
        name: quantity
        type: integer
      - in: formData
        name: seat_id
        type: integer
      - in: formData
        name: time
        type: string
      - in: formData
        name: title
        type: string
      - in: formData
        name: total_price
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controllers.Response'
            - properties:
                results:
                  $ref: '#/definitions/models.OrderBody'
              type: object
      tags:
      - Order Tiket
  /orders/cinema/{id}:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Add Choose Cinema tikers
      parameters:
      - description: Select Movie Tiket
        in: path
        name: id
        required: true
        type: integer
      - description: Search Name Cinema
        in: query
        name: searchName
        required: true
        type: string
      - description: Search Date Cinema
        in: query
        name: searchDate
        required: true
        type: string
      - description: Search Time Cinema
        in: query
        name: searchTime
        required: true
        type: string
      - description: Search Location Cinema
        in: query
        name: searchLocation
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controllers.Response'
            - properties:
                results:
                  $ref: '#/definitions/models.MoviesCinema'
              type: object
      tags:
      - Order Tiket
  /profile:
    get:
      consumes:
      - application/json
      description: Get Profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controllers.Response'
            - properties:
                results:
                  $ref: '#/definitions/models.PointProfile'
              type: object
      security:
      - ApiKeyAuth: []
      summary: Profile
      tags:
      - Profile
    patch:
      consumes:
      - multipart/form-data
      description: Update Movies
      parameters:
      - description: Update First Name
        in: formData
        name: First_Name
        required: true
        type: string
      - description: Update Last Name
        in: formData
        name: Last_Name
        required: true
        type: string
      - description: Update Image
        in: formData
        name: Image
        type: file
      - description: Update Phone_Number
        in: formData
        name: Phone_Number
        required: true
        type: string
      - description: Update Email
        in: formData
        name: Email
        required: true
        type: string
      - description: Update Password
        in: formData
        name: Password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controllers.Response'
            - properties:
                results:
                  $ref: '#/definitions/models.Profile'
              type: object
      tags:
      - Profile
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"